+++ learnflowserver/src/main/java/com/learnflow/learnflowserver/domain/Phase.java
@@ -1,5 +1,6 @@
-package com.learnflow.learnflowserver.entity;
+package com.learnflow.learnflowserver.domain;
 
+import com.learnflow.learnflowserver.domain.common.enums.Status;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
