+++ learnflowserver/src/main/java/com/learnflow/learnflowserver/domain/User.java
@@ -1,12 +1,11 @@
-package com.learnflow.learnflowserver.entity;
+package com.learnflow.learnflowserver.domain;
 
+import com.learnflow.learnflowserver.domain.common.enums.Role;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-
-import java.time.LocalDateTime;
 
 @Entity
 @Table(name = "users")
