+++ learnflowserver/src/main/java/com/learnflow/learnflowserver/domain/common/enums/Role.java
@@ -1,0 +1,5 @@
+package com.learnflow.learnflowserver.domain.common.enums;
+
+public enum Role {
+    STUDENT, TEACHER
+}
