+++ learnflowserver/src/main/java/com/learnflow/learnflowserver/domain/common/enums/Status.java
@@ -1,0 +1,5 @@
+package com.learnflow.learnflowserver.domain.common.enums;
+
+public enum Status {
+    ACTIVE, INACTIVE, COMPLETED
+}
