+++ learnflowserver/src/main/java/com/learnflow/learnflowserver/domain/common/enums/StudentStatus.java
@@ -1,4 +1,4 @@
-package com.learnflow.learnflowserver.entity;
+package com.learnflow.learnflowserver.domain.common.enums;
 
 public enum StudentStatus {
     NOT_STARTED, IN_PROGRESS, COMPLETED
