+++ learnflowserver/src/main/java/com/learnflow/learnflowserver/dto/PhaseInfoDto.java
@@ -1,6 +1,6 @@
 package com.learnflow.learnflowserver.dto;
 
-import com.learnflow.learnflowserver.entity.Status;
+import com.learnflow.learnflowserver.domain.common.enums.Status;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
