--- learnflowserver/src/main/java/com/learnflow/learnflowserver/entity/Role.java
@@ -1,5 +1,0 @@
-package com.learnflow.learnflowserver.entity;
-
-public enum Role {
-    STUDENT, TEACHER
-}
