--- learnflowserver/src/main/java/com/learnflow/learnflowserver/entity/Status.java
@@ -1,5 +1,0 @@
-package com.learnflow.learnflowserver.entity;
-
-public enum Status {
-    ACTIVE, INACTIVE, COMPLETED
-}
