+++ learnflowserver/src/main/java/com/learnflow/learnflowserver/repository/AssignmentRepository.java
@@ -1,7 +1,7 @@
 package com.learnflow.learnflowserver.repository;
 
-import com.learnflow.learnflowserver.entity.Assignment;
-import com.learnflow.learnflowserver.entity.User;
+import com.learnflow.learnflowserver.domain.Assignment;
+import com.learnflow.learnflowserver.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
